Reorganize to avoid circular imports
