+++ packages/syft/src/syft/service/file_object/file_service.py
@@ -7,14 +7,14 @@ from ...serde.serializable import serializable
 from ...store.document_store import DocumentStore
 from ...store.file_store import SyftResource
 from ...store.file_store import SyftWriteResource
+from ...types.file_object import CreateFileObject
+from ...types.file_object import FileObject
 from ...types.uid import UID
 from ..context import AuthedServiceContext
 from ..response import SyftError
 from ..service import AbstractService
 from ..service import TYPE_TO_SERVICE
 from ..service import service_method
-from .file_object import CreateFileObject
-from .file_object import FileObject
 from .file_stash import FileObjectStash
