+++ packages/syft/src/syft/service/file_object/file_stash.py
@@ -3,7 +3,7 @@ from ...serde.serializable import serializable
 from ...store.document_store import BaseUIDStoreStash
 from ...store.document_store import DocumentStore
 from ...store.document_store import PartitionSettings
-from .file_object import FileObject
+from ...types.file_object import FileObject
 
 
 @serializable()
